requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
boto3==1.35.89
botocore==1.35.89
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.41.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.5
Django==5.1.5
django-celery-beat==2.7.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.6.0
django-filter==25.1
django-rest-auth==0.9.5
django-storages==1.14.4
django-timezone-field==7.0
djangorestframework==3.15.2
drf-yasg==1.21.8
gunicorn==21.2.0
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.5
kombu==5.4.2
Markdown==3.7
MarkupSafe==3.0.2
packaging==24.2
pillow==11.0.0
pluggy==1.5.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
Pygments==2.18.0
pytest==8.3.4
pytest-django==4.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.0.1
requests==2.32.3
s3transfer==0.10.4
six==1.17.0
sqlparse==0.5.3
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
whitenoise[brotli]==6.6.0
